Changed drawing so that bottom row is fully visible.
As of now the top left corner is always aligned with the tiles. This is
what causes the problem in the first place. The ideal solution would be
to align the window with the bottom left tiles if you tried to center the
window on a tile closer than (screen_tiles_height/2 -1) to the south pole.

But, for now, I just grepped for occurrences where the ysize (or the
values derived from it) were used, and those places that had relevance
to drawing the map, and I added 1 (using the EXTRA_BOTTOM_ROW constant).

Patch submitted by Thue Janus Kristensen <thue@diku.dk>,
with Xaw implementation by me.


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@2152 a0f10bec-cc02-0410-94fc-a9cfff90b4cd