Added Nuclear Fallout.  Controlled by game.ruleset value nuke_contamination.
Industrialization and population still generate Pollution.  Dropping a Nuke
generates Nuclear Fallout, which is distinct from Pollution.  There is a
new command to clean Fallout vs. cleaning Pollution.

Similar to Pollution contributing to Global Warming, Fallout contributes to
Nuclear Winter -- which also changes terrain, but tends to Desert, Tundra
and Glacier.  Added a new "cooling" icon to the info area to indicate the
progress towards Nuclear Winter, and also icons for Fallout and Cleaning
Fallout for the main map.

GTK+ and Xaw clients work; made some changes to MUI client, but they are
probably incomplete and are untested.

Finally, AIs are more aggressive at cleaning up Pollution, but not Fallout.


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@2563 a0f10bec-cc02-0410-94fc-a9cfff90b4cd