Adds new logging macros and levels to the AI code. Use the CITY_LOG, UNIT_LOG, GOTO_LOG and BODYGUARD_LOG macros for these respective uses. Set log levels in ai/aitools.h and recompile to watch a specific part of the AI code.
Note that GOTO_LOG can be used where the unit parameter may be NULL or even a
dangling pointer, as long as the result parameter is properly set. It will only log
a situation of it is anomalous.

Also adds extra checking to autoexplorer code and fixes bug that caused
autoexplorers not go back into a city to bolster defenses when done exploring.


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@5035 a0f10bec-cc02-0410-94fc-a9cfff90b4cd