Change the format of the scorelog. Rewrite the scorelog generation.
This change fixes three known shortcomings of the old format:
 - there is no way to add players during a game
 - there is no way to match a scorelog to a savegame
 - the format uses years instead of turns

The first is addressed by adding information when a player is added
to the game and when a player is removed from the game.

The second is addressed by adding an id to the savegame and to the
scorelog. This id is generated at the start of the game. The id of the
scorelog and the id of the savegame must match if you want to append
data to a scorelog.

The last issues is easy. Just replace the years with turns. The year
information however is also given.

For a description of the new format visit doc/scorelog-v2.


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@5348 a0f10bec-cc02-0410-94fc-a9cfff90b4cd