Implement a function get_tile_boundaries that returns the corners of a boundary of a tile, and tells whether this boundary actually exists and is on the tile.
This simplifies the logic of drawing the map grid, national borders,
coastlines, city selection rectangles, and unavailable city tiles.  Several
buglets are also fixed: for instance the map grid used to be drawn for the
citydlg in iso-view but not in non-iso view.

Patch by me in PR#8507, with additional comments.


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@7468 a0f10bec-cc02-0410-94fc-a9cfff90b4cd