Change calculation of research cost. tech_cost_style 1 is changed from (researchcost*number of parents) to MAX((1+parents) * (researchcost/2) * sqrt(1+parents), researchcost), and is set as default.
Note: This exposes two unrelated problems. The first is a sometimes bug
that makes gtk2 science dialog bar fail to update. The second is that it
is now even more plain how ridiculously powerful Republic is (although
Republic has not changed as such).

Only gtk2 client updated in this commit.


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@7955 a0f10bec-cc02-0410-94fc-a9cfff90b4cd